made requested changes, added flow to home screen
