+++ app/src/main/AndroidManifest.xml
@@ -20,6 +20,7 @@
         </activity>
 
         <activity android:name=".ui.home.HomeActivity"/>
+        <activity android:name=".ui.login.LoginActivity"/>
 
     </application>
