+++ app/src/main/java/com/mealsmadeeasy/ui/splash/SplashActivity.kt
@@ -1,122 +1,14 @@
 package com.mealsmadeeasy.ui.splash
 
 import android.os.Bundle
-import android.content.Intent
-import android.support.v4.app.FragmentActivity
-import android.util.Log
 import com.mealsmadeeasy.ui.BaseActivity
-import com.mealsmadeeasy.ui.home.HomeActivity
-import com.mealsmadeeasy.R
-import android.view.View
+import com.mealsmadeeasy.ui.login.LoginActivity
 
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import com.google.android.gms.auth.api.signin.GoogleSignInClient
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-
-import com.google.android.gms.common.api.ApiException
-import com.google.android.gms.tasks.OnCompleteListener
-import com.google.android.gms.tasks.Task
-
-import com.google.firebase.auth.AuthCredential
-import com.google.firebase.auth.AuthResult
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.auth.GoogleAuthProvider
-import android.support.design.widget.Snackbar
-import android.widget.Button
-import com.google.android.gms.common.SignInButton
-import kotlinx.android.synthetic.main.activity_splash.*
-
-
-class SplashActivity : BaseActivity(), View.OnClickListener {
-
-    private val TAG = "SplashActivity"
-    private val RC_SIGN_IN = 9001
-
-    private var mAuth: FirebaseAuth? = null
-    private var mGoogleSignInClient: GoogleSignInClient? = null
-
+class SplashActivity : BaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-        findViewById<SignInButton>(R.id.sign_in_button).setOnClickListener(this)
-
-        // Configure sign-in to request the user's ID, email address, and basic
-        // profile. ID and basic profile are included in DEFAULT_SIGN_IN.
-        val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.default_web_client_id)) //OAuth 2.0 client ID
-                .requestEmail()
-                .build()
-
-        // Build a GoogleSignInClient with the options specified by gso.
-        mGoogleSignInClient = GoogleSignIn.getClient(this, gso)
-        mAuth = FirebaseAuth.getInstance()
-    }
-
-    public override fun onStart() {
-        super.onStart()
-        // Check if user is signed in (non-null) and update UI accordingly.
-        val currentUser = mAuth?.getCurrentUser()
-        updateUI(currentUser)
-    }
-
-    private fun signIn() {
-        val signInIntent = mGoogleSignInClient?.getSignInIntent()
-        startActivityForResult(signInIntent, RC_SIGN_IN)
-    }
-
-    public override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent) {
-        super.onActivityResult(requestCode, resultCode, data)
-
-        // Result returned from launching the Intent from GoogleSignInApi.getSignInIntent(...);
-        if (requestCode == RC_SIGN_IN) {
-            val task = GoogleSignIn.getSignedInAccountFromIntent(data)
-            try {
-                // Google Sign In was successful, authenticate with Firebase
-                val account = task.getResult(ApiException::class.java)
-                firebaseAuthWithGoogle(account)
-            } catch (e: ApiException) {
-                // Google Sign In failed, update UI appropriately
-                Log.w(TAG, "Google sign in failed", e)
-                updateUI(null)
-            }
-
-        }
-    }
-
-    private fun firebaseAuthWithGoogle(acct: GoogleSignInAccount) {
-        Log.d(TAG, "firebaseAuthWithGoogle:" + acct.id!!)
-
-        val credential = GoogleAuthProvider.getCredential(acct.idToken, null)
-        mAuth?.signInWithCredential(credential)
-                ?.addOnCompleteListener(this) { task ->
-                    if (task.isSuccessful) {
-                        // Sign in success
-                        Log.d(TAG, "signInWithCredential:success")
-                        val user = mAuth?.getCurrentUser()
-                        updateUI(user)
-                    } else {
-                        // If sign in fails, display a message to the user.
-                        Log.w(TAG, "signInWithCredential:failure", task.exception)
-                        Snackbar.make(findViewById(R.id.activity_splash), "Authentication Failed.", Snackbar.LENGTH_SHORT).show()
-                        updateUI(null)
-                    }
-                }
-    }
-
-    private fun updateUI(user: FirebaseUser?) {
-        if (user != null) {
-            findViewById<SignInButton>(R.id.sign_in_button).setVisibility(View.GONE)
-        } else {
-            findViewById<SignInButton>(R.id.sign_in_button).setVisibility(View.VISIBLE)
-        }
-    }
-
-    override fun onClick(v: View) {
-        when (v.getId()) {
-            R.id.sign_in_button -> signIn()
-        }
+        startActivity(LoginActivity.newIntent(this))
+        finish()
     }
 }
