+++ app/src/main/res/values/dimens.xml
@@ -12,4 +12,9 @@
     <dimen name="week_at_a_glance_meal_period_title_size">18sp</dimen>
     <dimen name="week_at_a_glance_meal_period_title_padding">8dp</dimen>
     <dimen name="week_at_a_glance_meal_padding">8dp</dimen>
+
+    <dimen name="login_icon_padding">60dp</dimen>
+    <dimen name="login_app_name_size">24sp</dimen>
+    <dimen name="login_button_padding_top">80dp</dimen>
+    <dimen name="login_button_padding_bottom">60dp</dimen>
 </resources>
